layout/style/test/chrome/test_stylesheet_clone_import_rule.html

<!DOCTYPE html>
<html lang="en-US">
<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>

<style>div { color: green; }</style>

<link id="theOnlyLink" rel="stylesheet" type="text/css" href="import_useless1.css">

<div id="theOnlyDiv">This text will change colors several times.</div>

<script>
  SimpleTest.waitForExplicitFinish();

  const Cu = SpecialPowers.Components.utils;
  const { ContentTaskUtils } = ChromeUtils.import("resource://testing-common/ContentTaskUtils.jsm");

  document.styleSheetChangeEventsEnabled = true;

  let theOnlyDiv = document.getElementById("theOnlyDiv");
  let stylesheet = document.getElementById("theOnlyLink").sheet;

  runTest().catch(function(reason) {
    ok(false, "Failed with reason: " + reason);
  }).then(function() {
    SimpleTest.finish();
  });

  function cssRulesToString(cssRules) {
    return Array.from(cssRules).map(rule => rule.cssText).join('');
  }

  async function runTest() {

    // Test that the div is initially red (from base.css)
    is(getComputedStyle(theOnlyDiv).color, "rgb(0, 128, 0)", "div begins as green.");

    // Create a Promise to watch for some number of events.
    let generateCountingListenerCallback = function(numberOfEventsToWaitFor, arrayToFill) {
      let eventsReceived = 0;
      return function(event) {
        info("Event processor: called with event " + event.rule.cssText);
        arrayToFill.push(event);
        return (++eventsReceived >= numberOfEventsToWaitFor);
      }
    };
    let events = [];
    let gotAllOurStyleRuleAddedEvents = ContentTaskUtils.waitForEvent(document,
      "StyleRuleAdded", true, generateCountingListenerCallback(2, events));

    // Insert some import rules.
    stylesheet.insertRule('@import url("import_useless2.css")', 0);
    stylesheet.insertRule('@import url("import_useless2.css")', 1);

    // Wait for the StyleRuleAdded events to be fired. This function returns the last event,
    // but we don't care what it contains -- only that the correct number of events fired.
    await gotAllOurStyleRuleAddedEvents;
    is(events.length, 2, "Got the expected number of StyleRuleAdded events.");

    // Do some sanity checking of our import rules.
    let primaryRules = stylesheet.cssRules;
    for (let i = 0; i < 2; ++i) {
      // Check that the main stylesheet rule is the same rule we saw in the event listener.
      is(primaryRules[i], events[i].rule, "Rule " + i + " matches the event listener.");

      // Get the imported sheet and confirm it is non-null.
      let importSheet = primaryRules[i].styleSheet;
      ok(importSheet, "Imported sheet " + i + " exists.");
      if (!importSheet) {
        info("Sheet " + i + " doesn't exist -- aborting.");
        return;
      }

      // Confirm the import sheet has rules.
      info("Retrieving rules for sheet " + i);
      let rules;

      try {
        rules = importSheet.cssRules;
      } catch (e) {
        throw "Exception thrown when retrieving rules for sheet " + i + ": " + e.message;
      }
      ok(rules, "Imported sheet " + i + " has rules.");
      if (!rules) {
        info("Rules for sheet " + i + " don't exist -- aborting.");
        return;
      }
    }

    // Make some helper variables for the comparison tests.
    let importSheet1 = primaryRules[0].styleSheet;
    let rules1 = importSheet1.cssRules;

    let importSheet2 = primaryRules[1].styleSheet;
    let rules2 = importSheet2.cssRules;

    // Confirm that these two sheets are meaningfully the same.
    is(cssRulesToString(rules1), cssRulesToString(rules2), "Cloned sheet rules are equivalent.");

    // Add a color-changing rule to the first stylesheet.
    importSheet1.insertRule('div { color: blue; }');
    rules1 = importSheet1.cssRules;

    // And make sure that it has an effect.
    is(getComputedStyle(theOnlyDiv).color, "rgb(0, 0, 255)", "div becomes blue.");

    // Make sure that the two sheets have different rules now.
    isnot(cssRulesToString(rules1), cssRulesToString(rules2), "Cloned sheet rules are no longer equivalent.");

    // Add a color-changing rule to the second stylesheet (that will mask the first).
    importSheet2.insertRule('div { color: red; }');
    // And make sure that it has an effect.
    is(getComputedStyle(theOnlyDiv).color, "rgb(255, 0, 0)", "div becomes red.");

    // Delete the second sheet by removing the import rule, and make sure the color changes back.
    stylesheet.deleteRule(1);
    is(getComputedStyle(theOnlyDiv).color, "rgb(0, 0, 255)", "div goes back to blue.");

    // Delete the first sheet by removing the import rule, and make sure the color changes back.
    stylesheet.deleteRule(0);
    is(getComputedStyle(theOnlyDiv).color, "rgb(0, 128, 0)", "div goes back to green.");
  }
</script>
</html>
